Reintegrated Firebase and moved beat.wav and TapInfo.js to proper locations
