+++ public/EdgeRACTrainer.js
