+++ public/engine/components/RectangleComponent.js
