+++ public/game/behaviours/CircleBehaviour.js
