+++ public/game/behaviours/ScoreCalculator.js
