+++ public/game/data/TapInfo.js
